+++ polishing/alignment.h
@@ -20,25 +20,25 @@ class Alignment
 public:
 	Alignment(size_t size, const SubstitutionMatrix& sm);
 
-	double globalAlignment(const std::string& v, const std::string& w, 
+	float globalAlignment(const std::string& v, const std::string& w, 
 						   int index);
-	double addDeletion(unsigned int index, unsigned int letterIndex);
-	double addSubstitution(unsigned int wordIndex, 
+	float addDeletion(unsigned int index, unsigned int letterIndex);
+	float addSubstitution(unsigned int wordIndex, 
 						   unsigned int letterIndex,
 						   char base, const std::string& read);
 
-	double addInsertion(unsigned int wordIndex,
-						unsigned int positionIndex,
-						char base, const std::string& read);
+	float addInsertion(unsigned int wordIndex,
+					   unsigned int positionIndex,
+					   char base, const std::string& read);
 
-	typedef Matrix<double> FloatMatrix;
+	typedef Matrix<float> FloatMatrix;
 private:
 	std::vector<FloatMatrix> _forwardScores;
 	std::vector<FloatMatrix> _reverseScores;
 	const SubstitutionMatrix& _subsMatrix;
 
-	double getBacktrackMatrix(const std::string& v, const std::string& w,
-							  FloatMatrix& scoreMat);
+	float getBacktrackMatrix(const std::string& v, const std::string& w,
+							 FloatMatrix& scoreMat);
 
 	void traceback(FloatMatrix& backtrack, const std::string& v, 
 				   const std::string& w, std::string& o_v,
